asdasdas

+++ database/migrations/2025_04_06_151220_add_user_id_to_notifications_table.php
@@ -8,14 +8,18 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('notifications', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')->after('id')->nullable()->index();
+            if (!Schema::hasColumn('notifications', 'user_id')) {
+                $table->foreignId('user_id')->nullable()->constrained()->onDelete('cascade');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('notifications', function (Blueprint $table) {
-            $table->dropColumn('user_id');
+            if (Schema::hasColumn('notifications', 'user_id')) {
+                $table->dropColumn('user_id');
+            }
         });
     }
 };
